Overhaul mesh organization
The mesh system for the scene has been restructured to use a base class
for meshes and mesh managers, to easily create new mesh classes by
deriving from the base.

Each mesh manager is related to a single pipeline type, and holds
multiple meshes of a specific type.

The scene and scene loader have been updated to reflect these changes,
and now allows billboards to be loaded from scene files.