Fix heap corruption caused by the debug messenger
The definition MTD_DEBUG may not be properly set in the
`DebugMessenger.hpp` header under certain conditions, causing
undesirable memory access in debug mode.

The debug callback function for the debug messenger has been changed to
a static function, since it should not be used by other regions of the
engine.