Create Model base class for mesh behavior
The `Model` class is the base for the application to implement custom
functionality for the meshes by extending this base class.

Derived classes must be registered to allow the engine to create
instances of the custom models.

The `EmptyModel` is used as a placeholder with no functionality when no
custom behavior is registered for the mesh.