Create event system
The event system allows both the engine and the application to create
and dispatch any kind of event, as well as set callbacks for these
events.

The events are stored in a queue, being handled before updating the
scene, also avoiding blocking event processing.

The `EventManager` stores the events and the callbacks, handling them.