Create frame profiler and its GUI
A profiler has been created to allow measuring the time taken for each
stage of the frame, as well as the whole frame.

Since the engine currently runs on a single thread, the sum of all frame
stages should equal the total frame time.

A GUI has been created to present a histogram with the timings of all
stages of the last frame.